fix: loop cap fires only under an armed sandbox budget — no more silent 1e8 truncation (#772) (#773)
The sandbox iteration cap's default arm (100,000,000) fired with no
sandbox configured: every 1e8 cumulative in-frame iterations, the running
loop silently exited early (__loop_exit__="limit", counter reset, exit 0).
Any program past 1e8 iterations computed "program + periodic truncation".

Minimal repro (before): i counting to 150000000 stops at exactly 100000000,
exit 0. Found by polymethod's full-vector brute-force differential — 4
corrupted outputs in 524,288 at n=19,m=1000, corrupted indices in the
arithmetic progression the cap predicts (spacing 1e8/1001); deterministic
and tier-independent (EIGS_JIT_OFF too).

The cap now fires only when g_sandbox_loop_max is explicitly armed
(sandbox_run max_iter — behavior unchanged, still covered by
test_vm_run_bytecode.eigs). All four sites share the fix: both out-of-line
helpers (also the JIT-thunk path via jit_helper_loop_{stall,cap}_check)
and both VM opcode bodies. EigsState.loop_iterations and
CallFrame.saved_loop_iter widen int -> long long: an uncapped frame can
exceed int range (signed-overflow UB); jit.c holds no reference to either
field, so no ABI impact.

Regression: suite [124] test_loop_cap_772.eigs runs a 1.05e8-iteration
loop to completion (~8s release). SPEC.md loop-termination passage
updated. Verified: release suite 3275/3275; ASan+UBSan detect_leaks=1
3279/3279, leak tally 0; polymethod differential re-run at n=19,m=1000
and n=16,m=4000: 0 mismatches (was 4 and 2).

Closes #772

+++ src/builtins.c
@@ -4495,7 +4495,7 @@ Value* builtin_sandbox_run(Value *arg) {
     val_decref(stub);
 
     int saved_max = g_sandbox_loop_max;
-    int saved_iters = g_loop_iterations;
+    long long saved_iters = g_loop_iterations;
     g_sandbox_loop_max = max_iter > 0 ? max_iter : 1000000;
     g_loop_iterations = 0;
     /* #292: arm the allocation budget. Save/restore so nested sandbox_run (or a
